refactor: it test to unit test.
